feat(api): снятие капа кадров движка (POST /fps) + живой fps и лимит в /state
Кап держит софтовый пацер, а не vsync: главный цикл кладёт в base + 0x1B206EC период кадра в единицах 3*мс (50 = 1/60 с в бою, 100 = 1/30 с в меню), пацер 0xB98070 после Present спит до истечения периода. Мод перезаписывает поле каждый кадр из render (внутри Present — после сеттера движка и до ожидания пацера): cap=off -> 0, fps=N -> 3000/N. Режим и действующий лимит видны в /state -> fps_cap (limit, не сырое поле: поле перезаписывается каждый кадр и его чтение давало гонку с сеттером).

Вместе с POST /dt {"fixed":true} (кадр = ровно 1/60 с симуляции) это ускоряет прогоны: один и тот же скрипт перелёта (RNG заморожен, тиковый триггер) — 20 FPS 20.3 с, 50 FPS 8.9 с, кап снят (~174 кадра/с) 5.4 с, max_y во всех трёх 32.66 м. /state.fps считается по кадрам кольцевого буфера за 500 мс (было среднее за сессию — при снятом капе показывало 40 вместо ~240). Клиент — drmod_api.fps_cap(); документация — docs/API.md §3.13 и QWEN.md «Frame pacer». Попутно убран лишний as usize в tas/hooks.rs (clippy).

+++ src/tas/hooks.rs
@@ -516,7 +516,7 @@ unsafe extern "thiscall" fn keyboard_poll_detour(this: *const u8) {
             if word & (1 << bit) == 0 {
                 continue;
             }
-            let dik = (word_index * 32 + bit) as usize;
+            let dik = word_index * 32 + bit;
             unsafe { *keys.add(dik) |= 0x80 };
         }
     }
